Fixed niche initialization bug
If a mod existed that contained all base game or all creation club plugins, but that mod had an ignore token.json file, that ignore token wouldn't be considered when deciding whether to auto-create the Base Game and CC Mod Settings. As a result, these plugins would be erroneously considered as "claimed", and the Base Game or CC ModSettings would (erroneously) not be auto-generated.